Merge-rebase correction.

+++ Assemblies/ProteanCMS/CMS/Admin/Admin.Redirects.vb
@@ -28,7 +28,8 @@ Partial Public Class Cms
                 moDbHelper = myWeb.moDbHelper
             End Sub
 
-            Public Function CreateRedirect(ByRef redirectType As String, ByRef OldUrl As String, ByRef NewUrl As String, Optional ByVal hiddenOldUrl As String = "", Optional ByVal pageId As Long = 0, Optional ByVal isParentPage As String = "") As String
+            Public Function CreateRedirect(ByRef redirectType As String, ByRef OldUrl As String, ByRef NewUrl As String, Optional ByVal hiddenOldUrl As String = "", Optional ByVal pageId As Integer = 0, Optional ByVal isParentPage As String = "") As String
+
                 Try
 
                     Dim rewriteXml As New XmlDocument
@@ -66,6 +67,7 @@ Partial Public Class Cms
                     'Determine all the paths that need to be redirected
                     ' If redirectType = "301Redirect" Then
                     If pageId > 0 Then
+
                         If isParentPage = "True" Then
                             Select Case redirectType
                                 Case "301Redirect"
@@ -91,6 +93,7 @@ Partial Public Class Cms
                             If matchString.StartsWith("/") Then
                                 matchString = matchString.TrimStart("/")
                             End If
+
                             folderRules.Add("Folder: " & OldUrl)
                             newRule.InnerXml = "<rule name=""Folder: " & OldUrl & """><match url=""^" & matchString & "(.*)""/><action type=""Redirect"" url=""" & NewUrl & "{R:1}"" /></rule>"
                             If CurrentRule Is Nothing Then
